+++ R/np.plot.engine.condbandwidth.R
@@ -256,6 +256,7 @@
     )
 
     if (surface.supported && perspective && !gradients){
+      .np_plot_progress_plan(1L)
       view = match.arg(view)
       rotate = (view == "rotate")
       
@@ -634,6 +635,8 @@
       quantreg.gradient.diagonal <- isTRUE(quantreg) && isTRUE(gradients)
       dsf <- if (gradients && !quantreg.gradient.diagonal) bws$xndim else 1L
       tot.dim = bws$xndim + bws$yndim - quantreg
+      .np_plot_progress_plan(if (!gradients) tot.dim else
+        sum(plot.gradient.available) * if (quantreg) 1L else tot.dim)
 
       gradient_component_index <- function(slice.index, component.index) {
         if (quantreg.gradient.diagonal) slice.index else component.index